chore: better structure in middlewares #14
