+++ server/app/middlewares/index.ts
@@ -1,5 +1,5 @@
 import auth from './auth.middleware';
-import verifyRegister from './verifyRegister';
+import verifyRegister from './verifyRegister.middleware';
 import verifyToken from './verifyToken.middleware';
 
 export default {
